The 'Serialization' folder is moved to the project root to avaid build issues on build agent.
