+++ dotnet/src/Functions/Functions.OpenAPI/Serialization/ArrayParameterValueSerializer.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Text.Json.Nodes;
 using System.Web;
 
-namespace Microsoft.SemanticKernel.Plugins.OpenApi.Builders.Serialization;
+namespace Microsoft.SemanticKernel.Plugins.OpenApi.Serialization;
 
 /// <summary>
 /// This class provides methods for serializing values of array parameters.
