+++ dotnet/src/Functions/Functions.OpenAPI/Serialization/SimpleStyleParameterSerializer.cs
@@ -4,7 +4,7 @@ using System;
 using System.Text.Json.Nodes;
 using Microsoft.SemanticKernel.Plugins.OpenApi.Model;
 
-namespace Microsoft.SemanticKernel.Plugins.OpenApi.Builders.Serialization;
+namespace Microsoft.SemanticKernel.Plugins.OpenApi.Serialization;
 
 /// <summary>
 /// Serializes REST API operation parameter of the 'Simple' style.
