+++ dotnet/src/Functions/Functions.OpenAPI/Serialization/SpaceDelimitedStyleParameterSerializer.cs
@@ -2,18 +2,17 @@
 
 using System;
 using System.Text.Json.Nodes;
-using System.Web;
 using Microsoft.SemanticKernel.Plugins.OpenApi.Model;
 
-namespace Microsoft.SemanticKernel.Plugins.OpenApi.Builders.Serialization;
+namespace Microsoft.SemanticKernel.Plugins.OpenApi.Serialization;
 
 /// <summary>
-/// Serializes REST API operation parameter of the 'Form' style.
+/// Serializes REST API operation parameter of the 'SpaceDelimited' style.
 /// </summary>
-internal static class FormStyleParameterSerializer
+internal static class SpaceDelimitedStyleParameterSerializer
 {
     /// <summary>
-    /// Serializes a REST API operation `Form` style parameter.
+    /// Serializes a REST API operation `SpaceDelimited` style parameter.
     /// </summary>
     /// <param name="parameter">The REST API operation parameter to serialize.</param>
     /// <param name="argument">The parameter argument.</param>
@@ -24,19 +23,17 @@ internal static class FormStyleParameterSerializer
 
         Verify.NotNull(parameter);
 
-        if (parameter.Style != RestApiOperationParameterStyle.Form)
+        if (parameter.Style != RestApiOperationParameterStyle.SpaceDelimited)
         {
-            throw new ArgumentException($"Unexpected Rest API operation parameter style - `{parameter.Style}`", nameof(parameter));
+            throw new ArgumentException($"Unexpected Rest API operation parameter style `{parameter.Style}`. Parameter name `{parameter.Name}`.", nameof(parameter));
         }
 
-        // Handling parameters of array type.
-        if (parameter.Type == ArrayType)
+        if (parameter.Type != ArrayType)
         {
-            return SerializeArrayParameter(parameter, argument);
+            throw new ArgumentException($"Serialization of Rest API operation parameters of type `{parameter.Type}` is not supported for the `{RestApiOperationParameterStyle.SpaceDelimited}` style parameters. Parameter name `{parameter.Name}`.", nameof(parameter));
         }
 
-        // Handling parameters of primitive - integer, string, etc type.
-        return $"{parameter.Name}={HttpUtility.UrlEncode(argument)}";
+        return SerializeArrayParameter(parameter, argument);
     }
 
     /// <summary>
@@ -49,7 +46,7 @@ internal static class FormStyleParameterSerializer
     {
         if (JsonNode.Parse(argument) is not JsonArray array)
         {
-            throw new KernelException($"Can't deserialize parameter name `{parameter.Name}` argument `{argument}` to JSON array");
+            throw new KernelException($"Can't deserialize parameter name `{parameter.Name}` argument `{argument}` to JSON array.");
         }
 
         if (parameter.Expand)
@@ -57,6 +54,6 @@ internal static class FormStyleParameterSerializer
             return ArrayParameterValueSerializer.SerializeArrayAsSeparateParameters(parameter.Name, array, delimiter: "&"); //id=1&id=2&id=3
         }
 
-        return $"{parameter.Name}={ArrayParameterValueSerializer.SerializeArrayAsDelimitedValues(array, delimiter: ",")}"; //id=1,2,3
+        return $"{parameter.Name}={ArrayParameterValueSerializer.SerializeArrayAsDelimitedValues(array, delimiter: "%20")}"; //id=1%202%203
     }
 }
