+++ dotnet/src/Functions/Functions.OpenApi/Builders/QueryStringBuilder.cs
@@ -3,8 +3,8 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Microsoft.SemanticKernel.Plugins.OpenApi.Builders.Serialization;
 using Microsoft.SemanticKernel.Plugins.OpenApi.Model;
+using Microsoft.SemanticKernel.Plugins.OpenApi.Serialization;
 
 namespace Microsoft.SemanticKernel.Plugins.OpenApi.Builders;
