+++ dotnet/src/Functions/Functions.OpenApi/Model/RestApiOperation.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
 using System.Text.RegularExpressions;
-using SimpleStyleParameterSerializer = Microsoft.SemanticKernel.Plugins.OpenApi.Builders.Serialization.SimpleStyleParameterSerializer;
+using Microsoft.SemanticKernel.Plugins.OpenApi.Serialization;
 
 namespace Microsoft.SemanticKernel.Plugins.OpenApi.Model;
