+++ dotnet/src/Functions/Functions.UnitTests/OpenApi/Builders/Serialization/SpaceDelimitedStyleParametersSerializerTests.cs
@@ -1,8 +1,8 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
 using System;
-using Microsoft.SemanticKernel.Plugins.OpenApi.Builders.Serialization;
 using Microsoft.SemanticKernel.Plugins.OpenApi.Model;
+using Microsoft.SemanticKernel.Plugins.OpenApi.Serialization;
 using Xunit;
 
 namespace SemanticKernel.Functions.UnitTests.OpenApi.Builders.Serialization;
